test: add note controller case (#939)

+++ test/helper/db-mock.helper copy.ts
@@ -1,5 +1,6 @@
 import { MongoMemoryServer } from 'mongodb-memory-server'
 import mongoose from 'mongoose'
+import { nanoid } from 'nanoid/async'
 
 import { getModelForClass } from '@typegoose/typegoose'
 import {
@@ -11,6 +12,7 @@ import {
 
 let mongod: MongoMemoryServer
 
+const dbMap = new Map<string, typeof mongoose>()
 /**
  
  * Connect to mock memory db.
@@ -19,25 +21,37 @@ const connect = async () => {
   mongod = await MongoMemoryServer.create()
   const uri = mongod.getUri()
 
-  return await mongoose.connect(uri, {
+  const mongooseInstance = await mongoose.connect(uri, {
     autoIndex: true,
     maxPoolSize: 10,
   })
+  const id = await nanoid()
+  dbMap.set(id, mongooseInstance)
+  return id
 }
 
 /**
  * Close db connection
  */
-const closeDatabase = async () => {
+const closeDatabase = async (id: string) => {
+  const mongoose = dbMap.get(id)
+  if (!mongoose) {
+    return
+  }
   await mongoose.connection.dropDatabase()
   await mongoose.connection.close()
-  await mongod.stop()
+  dbMap.delete(id)
+  if (dbMap.size === 0) await mongod.stop()
 }
 
 /**
  * Delete db collections
  */
-const clearDatabase = async () => {
+const clearDatabase = async (id: string) => {
+  const mongoose = dbMap.get(id)
+  if (!mongoose) {
+    return
+  }
   const collections = mongoose.connection.collections
 
   for (const key in collections) {
@@ -48,17 +62,13 @@ const clearDatabase = async () => {
 
 export const dbHelper = {
   connect,
-  close: closeDatabase,
-  clear: clearDatabase,
+  close: () => closeDatabase(),
+  clear: () => clearDatabase(),
 
   getModel<U extends AnyParamConstructor<any>, QueryHelpers = BeAnObject>(
     cl: U,
     options?: IModelOptions,
   ): ReturnModelType<U, QueryHelpers> {
-    return getModelForClass(cl, {
-      existingMongoose: mongoose,
-      existingConnection: mongoose.connection,
-      ...options,
-    })
+    return getModelForClass(cl, options)
   },
 }
